make column extensions normal functions
